requirements.txt
pandas
python-dotenv
langchain
langchain-community
sentence-transformers
faiss-cpu
transformers[agents]
langchain-huggingface